chore: move alloy conversions to alloy-compat.rs
